PM: Force Sleep results in hang
[Defect / Enhancement / New Feature]
MobC00386828

[Problem]
Suspend time is longer than 3.4.5 kernel. Due to this, once force
sleep command is issued, modem goes to retention, but AP is active.
When existing CP-AP interrupt is getting handled, AP writes to BINTC
register in modem domain which is now in LPM, and freezes.

[Solution]
As soon as force sleep command is issued, disable the CP-AP intr.
Skip checking for wakeup sources and initiate suspend immediately.